Pi/requirements.txt
mpu6050-gyroscope-accelerometer
pigpio
paramiko
